refactor(auth-lib): move fetchAuthTypeMetadata and DEFAULT_AUTH_TYPE_METADATA to svc.ts
